Pulling A Song (read description)
Can now pull a song, go to review page after logging in.
If its not working (it may say Spotify access token not found) go into the console and run: sessionStorage.getItem("spotifyAccessToken")
to see if the accessToken is being stored properly

RateSong.jsx added useEffect block to fetch a song and output the name and artist (on reviews page currently)

In UserInfo i moved useEffect block down to other spotify login related blocks (for exchanging Authorization code from URL for access token)

Created api/getSong/route.js for fetching songs.

Changed Redirect URI in .env from dashboard to UserInfo (The authorization code in the URL that we exchange for an Access Token is only present in the Redirect URI URL) (UserInfo was trying to obtain it from UserInfo page, but it was only present in Dashboard)