Now able to pull Recently Played Songs
api/getRecentlyPlayedSongs: GET function was missing REQ param. So we couldn't access the request headers to retrieve the access token

RateSong: Updated to set recentlyPlayed to data instead of data.items. Backend was already returning data.items so data in the frontend was actually the array of recently played songs.